// backend/3/extended-task-api/.adonisjs/client/registry/tree.d.ts
/* eslint-disable prettier/prettier */
import type { routes } from './index.ts'

export interface ApiDefinition {
  auth: {
    newAccount: {
      store: typeof routes['auth.new_account.store']
    }
    accessTokens: {
      store: typeof routes['auth.access_tokens.store']
    }
  }
  profile: {
    profile: {
      show: typeof routes['profile.profile.show']
    }
    accessTokens: {
      destroy: typeof routes['profile.access_tokens.destroy']
    }
    tasks: {
      showAll: typeof routes['profile.tasks.show_all']
      show: typeof routes['profile.tasks.show']
      store: typeof routes['profile.tasks.store']
      patch: typeof routes['profile.tasks.patch']
      delete: typeof routes['profile.tasks.delete']
    }
  }
}